modified:   commands/createProject.js 	modified:   commands/createcomment.js 	modified:   commands/delProject.js 	modified:   commands/delcomment.js 	modified:   commands/modcomment.js 	modified:   commands/modifyProject.js 	modified:   commands/mycomments.js 	modified:   commands/updateMilestone.js 	modified:   commands/viewGuildProjects.js 	modified:   commands/viewProject.js 	modified:   commands/viewcomments.js 	modified:   commands/viewmyprojects.js 	modified:   error_log.js

+++ error_log.js
@@ -1,24 +1,41 @@
 const fs = require('fs') 
       
-LogError = function(error) {
+error = function(error) {
     try {
-        const path = './logs/';
-        const base_name = 'BTB-log-';
-        const data = error;
+        const path = './logs';
+        const base_name = '/BTB-log-';
+        let data = error + "\n";
 
         const date = new Date();
 
         const year = date.getFullYear();
         const month = date.toLocaleString('default', { month: '2-digit' });
-        
-        const day = date.getDate();
-        const hours = date.getHours();
+        const hour = date.getHours();
         const minutes = date.getMinutes();
+        const day = date.getDate();
+        const seconds = date.getSeconds();
 
-        const time = year + "-" + month + "-" + day + " : " + hours + ":" +  "0" + minutes;
-        const file_name = base_name + time;
-        fs.writeFile(path + file_name, data, (err) => {
-            if (err) throw err;
+        const error_timestamp = year + "-" + month + "-" + day + " : " + hour + ":" + minutes + ":" + seconds;
+        data = error_timestamp + " | " + data;
+        const time = year + "-" + month + "-" + day;
+        const file_name = base_name + time + ".log";
+
+        fs.readdir(path, (err, files) => {
+            if (err) throw new Error('readdir');
+            let DoLogFileAlreadyExist = false;
+            files.forEach(file => {
+                if (file_name === "/" + file) {
+                    DoLogFileAlreadyExist = true;
+                    fs.appendFile(path + file_name, data, (err) => {
+                        if (err) throw new Error('appendFile');
+                    });
+                }
+            });
+            if (DoLogFileAlreadyExist === false) {
+                fs.writeFile(path + file_name, data, (err) => {
+                    if (err) throw err;
+                });
+            }
         });
     } catch (err) {
         console.log(err);
@@ -26,4 +43,4 @@ LogError = function(error) {
     }
 }
 
-LogError("This is a error");
+module.exports = {error}
